refactor/#262 폴더 위치 변경
